openvpp-agents/requirements.txt
aiomas==1.0.1
arrow==0.7.0
blosc==1.3.2
hg+ssh://hg@bitbucket.org/particon/chpsim@40be1747acbc851dc600ebde1555279bf7fde19c#egg=chpsim
click==6.6
coverage==4.0.3
Cython==0.24
docopt==0.6.2
flake8==2.5.4
h5py==2.5.0
mccabe==0.4.0
mosaik-api==2.2
msgpack-python==0.4.7
numpy==1.11.0
openvpp-agents==0.1.0
pep8==1.7.0
pluggy==0.3.1
psutil==4.1.0
py==1.4.31
pyflakes==1.1.0
pytest==2.9.1
pytest-asyncio==0.3.0
pytest-cov==2.2.1
python-dateutil==2.5.2
simpy==3.0.8
simpy.io==0.2.2
six==1.10.0
tox==2.3.1
virtualenv==15.0.1
